#352 pull out element wrapper utils
