+++ lib/taiko.js
@@ -15,6 +15,7 @@ const browserHandler = require('./handlers/browserHandler');
 const emulationHandler = require('./handlers/emulationHandler');
 const {match,$$,$$xpath,initElementSearch,isSelector,elements,element,isElementVisible} = require('./elementSearch');
 const { handleRelativeSearch, setDOM, RelativeSearchElement } = require('./proximityElementSearch');
+const {prepareParameters, getElementGetter, desc, initElementWrapperUtils} = require('./elementWrapperUtils');
 const { setConfig, defaultConfig, setNavigationOptions, setClickOptions, setBrowserOptions } = require('./config');
 const fs = require('fs');
 const path = require('path');
@@ -70,6 +71,7 @@ const connect_to_cri = async (target) => {
                 await Promise.all([runtime.enable(), network.enable(), page.enable(), dom.enable(), overlay.enable(), security.enable()]);
                 await setDOM(domHandler);
                 initElementSearch(runtimeHandler,handleRelativeSearch,descEvent);
+                initElementWrapperUtils(handleRelativeSearch,$$xpath);
                 if (defaultConfig.ignoreSSLErrors) security.setIgnoreCertificateErrors({ ignore: true });
                 _client.on('disconnect', reconnect);
                 device = process.env.TAIKO_EMULATE_DEVICE;
@@ -1533,57 +1535,6 @@ module.exports.$ = (selector, ...args) => {
     };
 };
 
-const prepareParameters = (attrValuePairs, options, ...args) => {
-    if (options instanceof RelativeSearchElement) {
-        args = [options].concat(args);
-        options = {};
-    }
-    let values = { selector:  { attrValuePairs: attrValuePairs, args: args } };
-    if (attrValuePairs instanceof RelativeSearchElement) {
-        args = [attrValuePairs].concat(args);
-        values.selector = { args: args};
-    } else  if (isString(attrValuePairs) || isSelector(attrValuePairs)) {
-        values.selector = { label: attrValuePairs, args: args };
-    }
-    values.options = options || {};
-    if(values.options.selectHiddenElement && hasProximitySelectors(values))
-        console.warn('WARNING: Proximity of hidden element are not available hence proximity selector will be ignored');
-    return values;
-};
-
-const getQuery = (attrValuePairs, tag = '') => {
-    let path = tag;
-    for (const key in attrValuePairs) {
-        if (key === 'class') path += `[contains(@${key}, ${xpath(attrValuePairs[key])})]`;
-        else path += `[@${key} = ${xpath(attrValuePairs[key])}]`;
-    }
-    return path;
-};
-
-
-const getElementGetter = (selector, query, tags, selectHiddenElement) => {
-    let get;
-    if (selector.attrValuePairs) {
-        let query = tags.split('|').map(tag => getQuery(selector.attrValuePairs, tag)).join('|');
-        get = async () => await handleRelativeSearch(await $$xpath(query, selectHiddenElement), selector.args);
-    }
-    else if (selector.label) get = async () => await handleRelativeSearch(await query(), selector.args);
-    else get = async () => await handleRelativeSearch(await $$xpath(tags, selectHiddenElement), selector.args);
-    return get;
-};
-
-const desc = (selector, query, tag) => {
-    let description = '';
-    if (selector.attrValuePairs) description = getQuery(selector.attrValuePairs, tag);
-    else if (selector.label) description = `${tag} with ${query} ${selector.label} `;
-
-    for (const arg of selector.args) {
-        description += description === '' ? tag : ' and';
-        description += ' ' + arg.toString();
-    }
-
-    return description;
-};
 
 /**
  * This {@link selector} lets you identify an image on a web page. This is done via the image's alt text or attribute and value pairs
@@ -1743,10 +1694,6 @@ module.exports.inputField = (attrValuePairs, ...args) => {
  * @returns {ElementWrapper}
  */
 module.exports.fileField = fileField;
-
-function hasProximitySelectors(values) {
-    return values.selector.args.length > 0;
-}
 
 function fileField(attrValuePairs, _options = {}, ...args) {
     validate();
